requirements.txt
Django==3.2.15
django-ninja==0.19.0
loguru==0.5.3
asgiref==3.4.1
pydantic==1.10.2
pymysql==1.0.2
django-cors-headers==3.11.0
email-validator==1.2.1
python-jose==3.3.0
orjson==3.6.8
cryptography==37.0.2
uvicorn[standard]==0.17.6
httpx==0.23.0
apscheduler==3.9.1
redis==4.3.1
tenacity==8.0.1
jinja2==3.1.2
django-redis==5.2.0
passlib==1.7.4
bcrypt==3.2.2
wait-for-it==2.2.1
